Implement metrics
This adds a Prometheus metrics endpoint for metrics collection.

The initial implementation provides three metrics: total IPs, latest IPs
and new IPs, same as the web interface. The method for updating the
metrics is currently just a goroutine firing every minute, which isn't
the best way of doing it. A later revision should move this into
(probably) the receive handler and only update when we receive new data.