Initial implementation of jobs
This allows logged-in users to request a specific scan to be run.

This is merely a queue and does not actively initiate any scans; instead
each scanning node should poll the /jobs endpoint for new jobs and
submit results with the job ID.

Part of this change "hides" the navbar result counts if not in the index
page; that is, no results are being displayed.