Store scan times as datetime
Storing as int / Unix time wasn't necessary as SQLite can natively
handle datetimes.

This also fixes an issue with storing as either Unix time or datetime
whereby the stored time would include the second, but the query from the
UI would be at 0 second, so most queries would fail. This change calls
time.Truncate() on the time prior to storing it.