Add a count for job results
As the results screen doesn't link entries to jobs, we had no way of
knowing how many results were found without manually searching for a
substring of an IP address.

This adds a count to the jobs view so you can see how many results were
returned for each job.