Implement a custom time type for jobs
Currently we just use time.Time (or NullTime) and return a fixed string.

To be able to return Unix time in metrics we need an actual time.Time
value.

This adds a jobTime type for accessing the Unix() method (via type
conversation back to time.Time) and a custom String() method for
returning an empty string inside the template.