Implement store and display of traceroutes
A feature request received from a network engineer was to display a
traceroute from the scan node to each IP address to help identify which
devices the firewall may need updating on.

This adds a traceroute table, a method for receiving traceroutes as
form data, and for displaying the traceroute. It also adds an icon next
to each result entry if there is an associated traceroute.