Move database handle to a global variable
Each read/write on the database was opening and closing it. This is
unnecessary and not recommended.

This change creates a global variable accessed by all functions which
access the database. The database is opened once at startup.

This could be further refactored to use custom types instead. For my own
later reference, these articles are useful:
https://medium.com/@benbjohnson/structuring-applications-in-go-3b04be4ff091
http://www.alexedwards.net/blog/organising-database-access