Make OAuth2 credentials file path a flag
This fixes the credentials file being hard-coded. The default value is
the same as the old hard-coded value; however it is now relative to the
data directory (which also defaults to the current directory).

Any relative path in the -credentials flag will be prepended with the
data directory. This is mostly to work around not being able to template
flag variables, but it makes sense anyway.