Refactor authentication and implement group auth
This splits the authHandler() function to store session information in a
struct and splits the logic into a couple of smaller functions.

Users can now be authenticated by group membership by adding a group
name to the new groups table. If a user doesn't exist in the users
table, group membership is checked. Group authentication only works for
G Suite domains (Basic or Enterprise) - it does not work for the old
Legacy accounts, or for Gmail accounts.