Bugfix: Auth silently fails with Admin SDK is not enabled
If trying to authenticate a user via a group and the Admin SDK API had
not been enabled on the GCP project, the API would return an
authentication error and not log why.

This now:

* Ensures the Groups API returned a 200
* If not, logs the error message

This also fixes the empty error handling condition when accessing the
database to log when a database error occurred.