Bugfix: scanData() should use txn for SELECTs
The scanData() function sets up some prepared statements, but the SELECT
was alone in reading outside of the transaction created. This sometimes
causes SQLite to return a "table is locked" error when trying to read
from the table.

This also fixes the return counter not being incremented for new
entries, although this hasn't been an issue as the value isn't consumed
anywhere in the application.