Replace Echo with Chi
This swaps out Echo, which is a full framework, for Chi, which is a
simple and fast router.

Echo is nice at first for having lots of features, such as structured
logging and automatic TLS, but the benefits of having all the bells and
whistles are outweighed by Echo not using the standard net/http handler
interface. This makes testing and using other standard middleware harder
than it needs to be.

Chi is a lightweight and very fast router/mux and provides most of the
features we actually need - routing, middleware support, including a
default logger. Other niceties such as auto TLS are easily replicated
using acme/autocert.

Chi uses the standard net/http interface for handlers, which will
facilitate writing tests.