Add some HTTP handler tests
Now that the HTTP handlers conform to the usual net/http handler
interface it's easier to test them.

This adds some tests for most of the handlers. For the ones which use
Chi's URL parameters we need to setup the router and a test server and
call it that way; all other handlers are invoked directly.