Bugfix: Make redirectHTTPS() actually work
Request.URL doesn't contain as much information as I expected, even
though the documentation clearly says that most of the fields aren't
populated.

This fixes the middleware by testing if Request.TLS is nil, and also
checking the error from net.SplitHostPort(), which errors if the address
doesn't contain a port.

We also weren't actually setting the host if the port was the standard
https port which led to an invalid Location: header being set.