Travis: Download dep binary instead of building it
Per the [dep
FAQ](https://golang.github.io/dep/docs/FAQ.html#how-do-i-use-dep-in-ci)
it's better to download the pre-built binary for dep rather than `go
get`ing it.

This also has changes to the build matrix to make it a bit more
straightforward and to allow setting the `DEP_VERSION` variable on all Go
versions easily. Previously the build matrix wouldn't include
`DEP_VERSION` from `env` because that version set its own `env`.

The `deploy` is now fixed to use Go 1.10, which will need to be manually
changed when Travis update the 1.x.