Redirect to the previous page after login
Previously you would always be redirected to / after login.

This changes it to redirect to the URI you're currently on (e.g. /job)
by storing the URI in a temporary session cookie, and destroying that
cookie after login.

The state and redirect cookies are updated to have a max 300 second
lifetime as they logins shouldn't take longer than that.