Display result numbers in the Job page or when 0 results
This adds the Total/Latest/New counters to the Job page as well as the
index as these are good for reference.

This also clearly shows 0 results and the search bar and new job buttons
when there are 0 results, which was a bug to not do so on the index
page.