Fix results tests
Struct literals shouldn't be used for scanData; there are quite a few
fields now so being explicit makes it more obvious and allows omitting
fields.

A job needs to be saved before posting job results, so add a saveJob()
call to the TestJobResultsHandler test.

Test submitting job results twice to ensure it fails the second time.