Use time.Time values for FirstSeen/LastSeen/Latest
Since storing of time values was changed in
9e214b9e8cd609fabf72858182499a14306aeab6 this broke the Latest link in
the UI, which used the String() value of scanTime (which assumed
stored times were truncated to time.Minute).

The idea behind using the stringified version of this was to allow
people to read and change the timestamp in the query string, but in
practice this isn't necessarily useful anyway.

The query string value for Latest (?lastseen=) now uses a Unix
timestamp. We also now use a scanTime value for FirstSeen and LastSeen
as the String() method already correctly formats to the same format.