Use go-bindata to embed views and static data
It's important to keep the views up-to-date when deploying a new binary.
The binary, views and static data are bundled into a tarball at the
moment but we can make this easier.

This change adds a `go generate` directive to use go-bindata to embed
all external files into the binary. This results in a larger binary file
(~20MB) with the convenience of easier deploys.

The Travis config is updated to install go-bindata, build the binary and
create a SHA256 checksum. The binary and checksum are deployed to
GitHub.