Fix Let's Encrypt http-01 challenge
I'm puzzled how this ever worked: Let's Encrypt http-01 challenges work
by hitting the webserver on HTTP to request a validation file.

Scan (somehow) used to be able to generate the cert without this
validation step happening, but that's stopped working.

This change adds the autocert handler into Chi's middleware list such
that if, on plain HTTP a Let's Encrypt challenge comes in, it's handled
properly to allow cert generation.