Bugfix: Update metrics directly in results handler
The total, latest and new metrics are updated after results are posted.

This was being done in the saveResults() function. The "latest" metric
is worked out based on the last submission time, but submission update
is handled in recvResults() after it has called saveResults(), so the
latest metric would always be wrong until the next submission.

This moves the metric update to the very last thing in the recvResults()
handler, after submission update,  so the values are updated correctly.