Travis: Move go generate below linters
golint is complaining about code generated by go-bindata in bindata.go.

It's not clear why this is happenening when a) this never used to
happen; and b) it doesn't happen locally for me.

Either way in Travis golint always fails, and we don't need to lint
bindata.go anyway, so this swaps the order of `go generate` and `go
lint`.  Generate needs to be before `go vet` because vet will fail if
the asset variables are missing.