Bugfix: Don't load auth config when auth disabled
The oauthConfig() function gets called early on and expects to find the
credentials file. This is done even if the `-no-auth` flag is specified
and causes the program to quit if the credentials file doesn't exist,
even though you disabled auth.

This wraps the oauthConfig() function in a check for the authDisabled.