UI: Don't display all results by default
The current behaviour of the UI is to load all results - even those
which are "gone" - and display them, unpaginated, in the browser.

For large result sets this is pretty slow to load. It's also not
necessary to see expired results. Users of the UI generally just want to
know what's happening now.

This replaces the "?active" query string with "?all" and flips the
defaults around. The ActiveOnly flag in the template is now AllResults.
The button behaviour is the same - labelled as "Active only" and
green/white depending on current state.