Add Makefile
It's not currently obvious for anyone except me how to properly build
the project, specifically generating the assets file as it's not
committed to the repo.

The Makefile will now automatically install go-bindata if it's not found
in the user's $GOPATH/bin, then run go generate when it thinks the
assets file is out of date.