Refactor database functions to a sqlite package
The existing application uses a global variable for the database
connection pool and HTTP handlers and database functions alike all share
this global variable.

The web application is now bound to an App struct containing a database
connection handle. All HTTP handlers are now methods of this struct and
read the DB handle from there.

There are a handful of HTTP handlers which directly access the database:
this is fine for now, but those should be updated to use a dedicated
method in the sqlite package.

Each database function is now a method of a sqlite.DB struct, which
itself embeds a *sql.DB.

This should make the app a bit tidier and allow further refactoring of
database code to use an interface to switch out the database if desired.

Note that this now can't be built Go < 1.13 as it uses the new
sql.NullTime. Previously we implemented this locally (as a copy from
lib/pq).