CircleCI: Split workflows again
The intention is that for non-master we only want to do a matrix build
on multiple Go versions, and not attempt any release to GitHub.

For master branch we don't need the matrix build, just the latest Go
version defined in the job. For non-tagged merges to master we want to
upload a pre-release to GitHub, and tagged releases have a release.

This is a bit verbose, but I think will resolve the issues I had after
trying in abafa865c9c7cf3113d12d9d49153dc9029ad6d6 where a magic tag
appeared in CircleCI (which I'm not sure where it came from) and it was
doing both upload jobs despite the tag filter on one. It seems like the
tag filter needs to be on all jobs in the workflow to be effective.
Doing that it's then clearer (but more verbose) to split into separate
workflows.