CircleCI: Apply filter on Branch bindata, ignore branch on Release
Filters weren't applied to the Branches bindata job, so ran even for
master. The filter has been moved up to that job to exclude it.

Ignore all branches on the Release workflow and just look at git tag.
It's not possible to filter by both branch and tag [0], which is unfortunate
as someone could tag something not merged and it would get deployed.
With the current setting both Pre-release and Release get run on merges
to master, and either ghr or GitHub are then creating a tag for the
release, which isn't desired.

[0] https://discuss.circleci.com/t/workflow-not-triggering-on-my-tag-filter/32218