CircleCI: Remove Pre-release workflow
Using ghr to create a pre-release in GitHub always creates a tag from
the name given to ghr. Merges to master were correctly only running the
Pre-release workflow, but then ghr would create a tag and so the Release
workflow would run on the tag.

We don't need the pre-releases, it just seemed like a nice-to-have.
Given we didn't have it with Travis either, it's now removed to simplify
things: the parameters are removed from the upload job (which is now
renamed to release) and to work around the fact that there's no easy way
to verify if a tag points to a commit merged to master, an approval job
is added in the Release workflow to have manual verification before
pushing binaries to GitHub.