Migrate static assets to embed.FS
HTML templates and static files were embedded in the binary using
go-bindata.

As of Go 1.16, the new embed package can perform the same thing natively
without needing extra tooling.

Everything which referenced a static asset now does so using an
embed.FS. This allowed simplifying quite a bit of the code handling that
as it's now much more straightforward to handle thanks to the FS
interface. The html/template package can parse embed.FS and net/http can
serve it straight as a FileServer.

The minimum Go version required now is 1.16, reflected in go.mod.

The Makefile is now no longer really needed, but the sample-data target
is useful, so it's been left in place along with a target for building
the binary, but the README now says to use go build instead of make.

Vendored dependencies are updated along with this.